services/speech-analysis/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
redis==5.0.1
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Speech-to-Text
deepgram-sdk==3.0.0
openai-whisper==20231117

# Audio Analysis
librosa==0.10.1
numpy==1.26.3
scipy==1.12.0
soundfile==0.12.1

# ML/AI
torch==2.1.2
torchaudio==2.1.2

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
